[Improve]  spark & flink app-id auto-increment Improvements
