+++ streampark-console/streampark-console-service/src/main/java/org/apache/streampark/console/core/entity/ApplicationLog.java
@@ -32,16 +32,28 @@ public class ApplicationLog {
 
     @TableId(type = IdType.AUTO)
     private Long id;
+
     /** appId */
     private Long appId;
-    /** applicationId */
-    private String yarnAppId;
+
+    /**
+     * 1: flink
+     * 2: spark
+     */
+    private Integer jobType;
+
+    /** clusterId */
+    private String clusterId;
+
     /** The address of the jobmanager, that is, the direct access address of the Flink web UI */
-    private String jobManagerUrl;
+    private String trackingUrl;
+
     /** start status */
     private Boolean success;
+
     /** option name */
     private Integer optionName;
+
     /** option time */
     private Date optionTime;
     /** exception at the start */
