+++ streampark-console/streampark-console-service/src/main/java/org/apache/streampark/console/core/mapper/ApplicationMapper.java
@@ -15,31 +15,12 @@
  * limitations under the License.
  */
 
-package org.apache.streampark.console.core.enums;
+package org.apache.streampark.console.core.mapper;
 
-import com.baomidou.mybatisplus.annotation.EnumValue;
-import lombok.Getter;
+import org.apache.streampark.console.core.entity.Application;
 
-import java.util.Arrays;
+import com.baomidou.mybatisplus.core.mapper.BaseMapper;
 
-/** Compute engine type. */
-@Getter
-public enum EngineTypeEnum {
+public interface ApplicationMapper extends BaseMapper<Application> {
 
-    /** Apache Flink: activated by default */
-    FLINK(1),
-
-    /** Apache Spark */
-    SPARK(2);
-
-    @EnumValue
-    private final int code;
-
-    EngineTypeEnum(int code) {
-        this.code = code;
-    }
-
-    public static EngineTypeEnum of(Integer code) {
-        return Arrays.stream(values()).filter((x) -> x.code == code).findFirst().orElse(null);
-    }
 }
