+++ streampark-console/streampark-console-service/src/main/java/org/apache/streampark/console/core/service/application/AppBuildPipeService.java
@@ -15,9 +15,10 @@
  * limitations under the License.
  */
 
-package org.apache.streampark.console.core.service;
+package org.apache.streampark.console.core.service.application;
 
 import org.apache.streampark.console.core.entity.AppBuildPipeline;
+import org.apache.streampark.flink.packer.pipeline.DockerResolvedSnapshot;
 import org.apache.streampark.flink.packer.pipeline.PipelineStatusEnum;
 
 import com.baomidou.mybatisplus.extension.service.IService;
@@ -29,7 +30,7 @@ import java.util.Map;
 import java.util.Optional;
 
 /** Applications can be built asynchronously, can manage pipeline and get info */
-public interface SparkAppBuildPipeService extends IService<AppBuildPipeline> {
+public interface AppBuildPipeService extends IService<AppBuildPipeline> {
 
     /**
      * Build application. This is an async call method.
@@ -47,6 +48,14 @@ public interface SparkAppBuildPipeService extends IService<AppBuildPipeline> {
      * @return ApplicationBuildPipeline instance
      */
     Optional<AppBuildPipeline> getCurrentBuildPipeline(@Nonnull Long appId);
+
+    /**
+     * Get Docker resolved snapshot of specified application.
+     *
+     * @param appId application id
+     * @return DockerResolvedSnapshot instance
+     */
+    DockerResolvedSnapshot getDockerProgressDetailSnapshot(@Nonnull Long appId);
 
     /**
      * Whether the application can currently start a new building progress
