+++ streampark-console/streampark-console-service/src/main/resources/db/schema-h2.sql
@@ -16,19 +16,14 @@
  */
 
 -- ----------------------------
--- Table structure for t_flink_app_backup
+-- Table structure for t_app
 -- ----------------------------
-create table if not exists `t_flink_app_backup` (
-  `id` bigint generated by default as identity not null,
-  `app_id` bigint default null,
-  `sql_id` bigint default null,
-  `config_id` bigint default null,
-  `version` int default null,
-  `path` varchar(128)  default null,
-  `description` varchar(255) default null,
-  `create_time` datetime default null comment 'create time',
-  `modify_time` datetime default null comment 'modify time',
-  primary key(`id`)
+create table if not exists `t_app` (
+`id` bigint generated by default as identity not null,
+`job_type` tinyint default null,
+`create_time` datetime default null comment 'create time',
+`modify_time` datetime default null comment 'modify time',
+primary key(`id`)
 );
 
 -- ----------------------------
@@ -98,6 +93,24 @@ create table if not exists `t_flink_app` (
   `hadoop_user` varchar(500) default null,
   primary key(`id`)
 );
+
+
+-- ----------------------------
+-- Table structure for t_flink_app_backup
+-- ----------------------------
+create table if not exists `t_flink_app_backup` (
+`id` bigint generated by default as identity not null,
+`app_id` bigint default null,
+`sql_id` bigint default null,
+`config_id` bigint default null,
+`version` int default null,
+`path` varchar(128)  default null,
+`description` varchar(255) default null,
+`create_time` datetime default null comment 'create time',
+`modify_time` datetime default null comment 'modify time',
+primary key(`id`)
+);
+
 
 -- ----------------------------
 -- Table structure for t_flink_config
